+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Improvements
 - Allow style properties to set style values with a higher priority, which allows elements to style their default children
 - Allow changing the entire ui style for a single element
 - Skip unnecessary area updates for elements with dirty parents
+- Calculate panel scroll bar height based on content height
 
 Fixes
 - Fixed VerticalSpace height parameter being an integer